requirements.txt
Flask==2.3.3
pywebview[qt]==4.4.1
python-escpos==3.1
Flask-Cors==4.0.1
python-barcode==0.15.1